Improve unit tests
